+++ src/brevitas_examples/llm/main.py
@@ -11,14 +11,14 @@ import pprint
 import sys
 import warnings
 
+from brevitas_example.common.dynamo_utils import dynamo_export_ctx
+from brevitas_example.common.dynamo_utils import patch_dynamo_export
 import numpy as np
-from packaging import version
 import torch
 from torch.utils.data import DataLoader
 from transformers import AutoModelForCausalLM
 from transformers import AutoTokenizer
 
-from brevitas import torch_version
 from brevitas.export.inference.manager import quant_inference_mode
 from brevitas.export.onnx.standard.qcdq.manager import StdQCDQONNXManager
 from brevitas.graph import load_quant_model_mode
@@ -84,46 +84,8 @@ except:
     SharkManager = None
     logging.debug("Shark-AI not installed, cannot export to Shark")
 
-
-def _maybe_patch_dynamo_export():
-    # torch._dynamo.export in torch 2.10 and 2.11 can crash with
-    # "'NoneType' object has no attribute 'is_tensor'" when a raw None is left on
-    # Dynamo's symbolic stack: OutputGraph.compile_subgraph calls x.is_tensor()
-    # over those stack values. Upstream fixed this in 2.12 by replacing raw None
-    # on the stack with ConstantVariable(None) (pytorch/pytorch#169325). We
-    # backport that behaviour out-of-source for 2.10/2.11 only by coercing raw
-    # None to a ConstantVariable when the stack values are gathered.
-    if not (version.parse('2.10') <= torch_version < version.parse('2.12')):
-        return
-    from torch._dynamo.output_graph import OutputGraph
-    from torch._dynamo.variables import ConstantVariable
-    if getattr(OutputGraph._get_stack_values_to_restore, '_brevitas_none_patch', False):
-        return
-    original_fn = OutputGraph._get_stack_values_to_restore
-
-    @functools.wraps(original_fn)
-    def _get_stack_values_to_restore(self, tx, stack_pops):
-        stack_values, meta = original_fn(self, tx, stack_pops)
-        stack_values = [ConstantVariable.create(None) if v is None else v for v in stack_values]
-        return stack_values, meta
-
-    _get_stack_values_to_restore._brevitas_none_patch = True
-    OutputGraph._get_stack_values_to_restore = _get_stack_values_to_restore
-
-
-_maybe_patch_dynamo_export()
-
-
-def dynamo_export_ctx():
-    # From torch 2.10 onwards, torch._dynamo.export inlines built-in nn modules
-    # (install_free_tensors_for_export=True) instead of emitting call_module
-    # nodes. Setting install_free_tensors_for_export=False routes them back
-    # through the specialized NNModuleVariable path, restoring the pre-2.10 graph
-    # structure. The flag does not exist before torch 2.10.
-    if torch_version >= version.parse('2.10'):
-        import torch._dynamo.config as dynamo_config
-        return dynamo_config.patch(install_free_tensors_for_export=False)
-    return nullcontext()
+# We need to patch this before anything else is executed
+patch_dynamo_export()
 
 
 def filter_results(results, tasks):
